pyproject.toml
[project]
name = "etl-for-dumdums"
version = "0.1.0"
description = "ETL jobs for syncing data from various APIs to BigQuery"
requires-python = ">=3.11"
dependencies = [
    "dbt-bigquery>=1.9.0",
    "google-cloud-bigquery>=3.0.0",
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "playwright>=1.56.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "pytrends>=4.9.0",
    "requests>=2.28.0",
    "streamlit>=1.29.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
]

[tool.uv.sources]
etl-for-dumdums = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "sources"]

[tool.pyright]
extraPaths = ["."]

[tool.ruff.lint]
ignore = ["E402"]
